Comment FIXMEs about OUTER JOINs
Because Django doesn't do them, we beat the database
to hell foreach prkl.

We can't get the username or, as related, the upcoming
can_vote stuff, without hitting the database and querying
separately.

There's a very real chance this has to be made a custom
sql, unless a convenient hack can be conjured for outer
joins...

I think one-step outer joins should be simple-ish,
like putting None as the value in a .extra() dictionary.