Real voting
Probably must do more transaction management on the prkl object side.

It just commits after incr/decr, but afaict it uses a different cursor
on the same connection as the vote stuff, so it might go weird. Maybe.

GIT AMEND HERE:
raise ValueError('foo') caused the incr to happen but no vote to be
committed. This is not good, this is bad.