Recreate hash for non-existing cookies and force bootstrap
What that means is, if a cookie does not exist in the database,
calculate a new local hash for it.

More importantly, because we're checking voting abilities based on
cookies, and we may not have a cookie, we need to forge it.

However, the output stage does not set a cookie in the response
if one was received in the input stage WHERE WE FAKE IT.

Therefore a boolean is passed in request.META to tell if this
was fake or not.

Not beautiful but it works...