Custom settings and csrf middleware from Django's
This sets the session key as '' in cookies if a session
is not found. Mostly used when the site loads and the request
doesn't know whether or not it set the cookies correctly.

The csrf middleware compares to this empty-session key if
the Django csrf middleware returns forbidden.

What this means is every first page load will have the
same csrf token, which is very unsafe, but good enough for now.